// design-mockup/src/lib/types/concrete/MnKappaRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DesignSituation } from "./DesignSituation";
import type { ReinforcementCage } from "./ReinforcementCage";
import type { SteelBranch } from "./SteelBranch";

/**
 * Verzoek om het M-N-κ-diagram van een korf, los van een staaf.
 */
export type MnKappaRequest = { width_mm: number, height_mm: number, concrete_class: string, reinforcement_grade: string, cage: ReinforcementCage, 
/**
 * Normaalkracht waarbij het M-κ-diagram wordt bepaald (kN, trek positief).
 */
n_ed_kn: number, 
/**
 * Richting van het moment: +1 trek onder (standaard), −1 trek boven.
 */
moment_sign: number, n_strips: number, steel_branch: SteelBranch, design_situation: DesignSituation, 
/**
 * Aantal punten van het N-M-interactiediagram (0 = niet berekenen).
 */
interaction_points: number, };